feat: invlaidate cache based on headers, query and body
