+++ packages/repack/babel.config.js
@@ -23,4 +23,18 @@ module.exports = {
       ],
     },
   ],
+  env: {
+    test: {
+      presets: [
+        [
+          '@babel/preset-env',
+          {
+            targets: {
+              node: 12,
+            },
+          },
+        ],
+      ],
+    },
+  },
 };
