+++ packages/repack/package.json
@@ -78,6 +78,7 @@
     "ora": "^5.3.0",
     "pretty-format": "^26.6.2",
     "schema-utils": "^3.0.0",
+    "shallowequal": "^1.1.0",
     "source-map": "^0.7.3",
     "split2": "^3.2.2",
     "tapable": "^2.1.1",
@@ -101,6 +102,7 @@
     "@types/jest": "^26.0.21",
     "@types/node": "^14.14.28",
     "@types/react-dom": "^17.0.7",
+    "@types/shallowequal": "^1.1.1",
     "@types/split2": "^2.1.6",
     "@types/ws": "^7.4.0",
     "babel-jest": "^26.6.3",
@@ -108,6 +110,7 @@
     "del-cli": "^3.0.1",
     "eslint": "^7.30.0",
     "jest": "^26.6.3",
+    "node-fetch": "^3.0.0",
     "react-native": "^0.64.1",
     "release-it": "^14.10.0",
     "terser-webpack-plugin": "^5.1.3",
