+++ packages/repack/src/client/api/__tests__/ChunkManagerBackend.spec.ts
@@ -1,0 +1,228 @@
+import { ChunkManagerBackend } from '../ChunkManagerBackend';
+
+class FakeCache {
+  data: Record<string, string> = {};
+
+  async setItem(key: string, value: string) {
+    this.data[key] = value;
+  }
+
+  async getItem(key: string) {
+    return this.data[key] ?? null;
+  }
+
+  async removeItem(key: string) {
+    delete this.data[key];
+  }
+}
+
+describe('ChunkManager', () => {
+  it('should resolve with url only', async () => {
+    const manager = new ChunkManagerBackend({});
+    const cache = new FakeCache();
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+        };
+      },
+    });
+
+    const config = await manager.resolveChunk('src_App_js', 'main');
+    expect(config).toEqual({
+      url: 'http://domain.ext/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'GET',
+    });
+
+    const { fetch } = await manager.resolveChunk('src_App_js', 'main');
+    expect(fetch).toBe(false);
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/subpath/${chunkId}`,
+        };
+      },
+    });
+
+    const newConfig = await manager.resolveChunk('src_App_js', 'main');
+    expect(newConfig).toEqual({
+      url: 'http://domain.ext/subpath/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'GET',
+    });
+  });
+
+  it('should resolve with custom extension', async () => {
+    const manager = new ChunkManagerBackend({});
+    const cache = new FakeCache();
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}.js`,
+          excludeExtension: true,
+        };
+      },
+    });
+
+    const config = await manager.resolveChunk('src_App_js', 'main');
+    expect(config).toEqual({
+      url: 'http://domain.ext/src_App_js.js',
+      fetch: true,
+      method: 'GET',
+    });
+  });
+
+  it('should resolve with query', async () => {
+    const manager = new ChunkManagerBackend({});
+    const cache = new FakeCache();
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+          query: {
+            accessCode: '1234',
+            accessUid: 'asdf',
+          },
+        };
+      },
+    });
+
+    const config = await manager.resolveChunk('src_App_js', 'main');
+    expect(config).toEqual({
+      url: 'http://domain.ext/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'GET',
+      query: 'accessCode=1234&accessUid=asdf',
+    });
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+          query: 'token=some_token',
+        };
+      },
+    });
+
+    const { fetch, query } = await manager.resolveChunk('src_App_js', 'main');
+    expect(fetch).toBe(true);
+    expect(query).toEqual('token=some_token');
+  });
+
+  it('should resolve with headers', async () => {
+    const manager = new ChunkManagerBackend({});
+    const cache = new FakeCache();
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+          headers: {
+            'x-hello': 'world',
+          },
+        };
+      },
+    });
+
+    const config = await manager.resolveChunk('src_App_js', 'main');
+    expect(config).toEqual({
+      url: 'http://domain.ext/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'GET',
+      headers: { 'x-hello': 'world' },
+    });
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+          headers: {
+            'x-hello': 'world',
+            'x-changed': 'true',
+          },
+        };
+      },
+    });
+
+    const newConfig = await manager.resolveChunk('src_App_js', 'main');
+    expect(newConfig).toEqual({
+      url: 'http://domain.ext/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'GET',
+      headers: { 'x-hello': 'world', 'x-changed': 'true' },
+    });
+  });
+
+  it('should resolve with body', async () => {
+    const manager = new ChunkManagerBackend({});
+    const cache = new FakeCache();
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+          method: 'POST',
+          body: 'hello_world',
+        };
+      },
+    });
+
+    const config = await manager.resolveChunk('src_App_js', 'main');
+    expect(config).toEqual({
+      url: 'http://domain.ext/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'POST',
+      body: 'hello_world',
+    });
+
+    manager.configure({
+      storage: cache,
+      resolveRemoteChunk: async (chunkId, parentChunkId) => {
+        expect(parentChunkId).toEqual('main');
+
+        return {
+          url: `http://domain.ext/${chunkId}`,
+          body: 'message',
+          method: 'POST',
+        };
+      },
+    });
+
+    const newConfig = await manager.resolveChunk('src_App_js', 'main');
+    expect(newConfig).toEqual({
+      url: 'http://domain.ext/src_App_js.chunk.bundle',
+      fetch: true,
+      method: 'POST',
+      body: 'message',
+    });
+  });
+});
