+++ packages/repack/src/client/api/index.ts
@@ -1,1 +1,3 @@
 export * from './ChunkManager';
+export * from './Chunk';
+export * from './types';
